pyproject.toml
[project]
name = "ragwebsite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "chromadb>=1.0.13",
    "faiss-cpu>=1.11.0",
    "faster-whisper>=1.1.1",
    "google-generativeai>=0.8.5",
    "openai-agents>=0.1.0",
    "openai-whisper>=20250625",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.46.1",
    "yt-dlp>=2025.6.30",
]
